Move `form-control.ts` to form directory
