+++ libs/core/src/components/button/button.ts
@@ -14,7 +14,7 @@ import style from './button.style.css?inline'
 
 import { gdsCustomElement, html as customElementHtml } from '../../scoping'
 import { stripWhitespace } from '../../utils/helpers/strip-white-space'
-import { GdsFormControlElement } from '../../components/form-control'
+import { GdsFormControlElement } from '../form/form-control'
 import type { GdsElement } from '../../gds-element'
 
 // Create a customized `html` template tag that strips whitespace and applies custom element scoping.
